Multiple independent sets

This function estimates the SLR function using the multiple independent sets method. The kernel density estimate (KDE) of numerator and denominator density is computed on multiple independent sets and the estimated densities are averaged to compute the SLR. The independent sets used to estimate the SLR function are from KM_train and KNM_train. The function then produces various output based on KM_test, KNM_test, and unknown.

KM_train, KM_test, KNM_train, KNM_train should all be data frames containing columns named source1, source2, dep1, and dep2. Any remaining columns must be scores. There may be up to five scores.

Usage

multiple_kde(
  KM_train,
  KM_test,
  KNM_train,
  KNM_test,
  NUM_SETS = 10,
  unknown = NULL
)

Arguments

KM_train

known matches (KM) from training dataset.

KM_test

known matches (KM) from testing dataset.

KNM_train

known non matches (KNM) from training dataset.

KNM_test

known non matches (KNM) from testing dataset.

NUM_SETS

number of independent sets to use.

unknown

data frame of scores from an unknown case. Column names should match the variable names set for scores.

Value

A list containing the following components:

KM_SLR

SLRs for KM_test.

KNM_SLR

SLRs for KNM_test.

threshold

optimal threshold calculated using KM_train and KNM_train.

new_SLR

SLRs for unknown.

ROC_values

data frame containing columns the true positive rate tpr and false positive rate fpr computed using KM_test and KNM_test.

Examples

library(dplyr)
# Set up data
shoedata_split <- dep_split(shoedata, 0.75)
KM_train <- shoedata_split %>% filter(source1 == source2 & train == TRUE)
KM_test <- shoedata_split %>% filter(source1 == source2 & train == FALSE)
KNM_train <- shoedata_split %>% filter(source1 != source2 & train == TRUE)
KNM_test <- shoedata_split %>% filter(source1 != source2 & train == FALSE)
unknown <- data.frame(clique_size = c(5, 8), med_dist_euc = c(1.9, 1.1), 
                      input_overlap = c(.01, 0.26))

res <- multiple_kde(KM_train, KM_test, KNM_train, KNM_test, 5, unknown)