Strict independent set

This function estimates the SLR function using the strict independent set method, computing the numerator and denominator density using only an independent set from KM_train and KNM_train, respectively. The function then produces various output based on KM_test, KNM_test, and unknown.

KM_train, KM_test, KNM_train, KNM_train should all be data frames containing columns named source1, source2, dep1, and dep2. Any remaining columns must be scores. There may be up to five scores.

strict_indep_set(
  KM_train,
  KM_test,
  KNM_train,
  KNM_test,
  unknown = NULL,
  seed = NULL
)

Arguments

KM_train

known matches (KM) from training dataset.

KM_test

known matches (KM) from testing dataset.

KNM_train

known non matches (KNM) from training dataset.

KNM_test

known non matches (KNM) from testing dataset.

unknown

data frame of scores from an unknown case. Column names should match the variable names set for scores.

seed

seed for R's random number generator.

Value

A list containing the following components:

KM_SLR

SLRs for KM_test.

KNM_SLR

SLRs for KNM_test.

threshold

optimal threshold calculated using KM_train and KNM_train.

new_SLR

SLRs for unknown.

ROC_values

data frame containing columns the true positive rate tpr and false positive rate fpr computed using KM_test and KNM_test.

Examples

library(dplyr)
# Set up data
shoedata_split <- dep_split(shoedata, 0.75)
KM_train <- shoedata_split %>% filter(source1 == source2 & train == TRUE)
KM_test <- shoedata_split %>% filter(source1 == source2 & train == FALSE)
KNM_train <- shoedata_split %>% filter(source1 != source2 & train == TRUE)
KNM_test <- shoedata_split %>% filter(source1 != source2 & train == FALSE)
unknown <- data.frame(clique_size = c(5, 8), med_dist_euc = c(1.9, 1.1), 
                      input_overlap = c(.01, 0.26))

res <- strict_indep_set(KM_train, KM_test, KNM_train, KNM_test, unknown)
#> Warning: ROC_values not calculated0, 0