Results of one or more method of computing the SLR

This function returns the results of one or more method of computing the SLR using the method functions in the package.

slr_results(
  data,
  unknown = NULL,
  method = c("IgnoreDependence", "StrictIndependentSet", "AverageFeatures",
    "MultipleKDE"),
  NUM_SETS = 10,
  seed = NULL
)

Arguments

data

data frame containing columns named source1, source2, dep1, dep2, and train. The train column should be a logical vector with TRUE denoting the training set and FALSE the test set. Any remaining columns must be scores. There may be up to five scores.

unknown

data frame of scores from an unknown case. Column names should match the variable names set for scores.

method

vector of methods to calculate the SLR. Options include "IgnoreDependence", "StrictIndependentSet", "AverageFeatures", and "MultipleKDE".

NUM_SETS

number of sets to use for multiple KDE method.

seed

seed for R's random number generator.

Value

A list of the output for each of the methods in method. Each element of the list is a list containing the following components:

KM_SLR

SLRs for KM_test.

KNM_SLR

SLRs for KNM_test.

threshold

optimal threshold calculated using KM_train and KNM_train.

new_SLR

SLRs for unknown.

ROC_values

data frame containing columns the true positive rate tpr and false positive rate fpr computed using KM_test and KNM_test.

See also

Examples

# Set up
shoedata_split <- dep_split(shoedata, 0.75)
unknown <- data.frame(clique_size = c(5, 8), med_dist_euc = c(1.9, 1.1), 
                      input_overlap = c(.01, 0.26))

res_all <- slr_results(shoedata_split, unknown) # all methods
res_def <- slr_results(shoedata_split, unknown,
            c("IgnoreDependence", "StrictIndependentSet")) # default methods
res_prop <- slr_results(shoedata_split, unknown, 
            c("AverageFeatures", "MultipleKDE")) # proposed methods